Moved code into resourceTreeUtils file & changed action names
